fix:レンダー時の負荷を軽減するため、画像合成処理を移動

+++ scripts/menuThumbnailGen.js
@@ -2,34 +2,47 @@ const fs = require('fs');
 const path = require('path');
 
 // ファイルツリーを作成する再帰関数
-function createTree(directoryPath) {
+function createTree(directoryPath, newName) {
   const tree = {};
 
   const items = fs.readdirSync(directoryPath);
 
   for (const item of items) {
     const itemPath = path.join(directoryPath, item);
-    const newName = 'pathUrl'; // プロパティ名を変更する
     if (fs.statSync(itemPath).isDirectory()) {
-      tree[item] = createTree(itemPath, newName);
+      const name = newName || item; // プロパティ名を変更する
+      tree[name] = createTree(itemPath, newName);
     } else {
       const relativePath = path.relative(rootDirectory, itemPath);
-      tree[newName] = relativePath;
+      tree[newName || item] = relativePath;
     }
   }
 
   return tree;
 }
 
+// コマンドライン引数からディレクトリパスとアウトプットファイルパスを取得
+if (process.argv.length < 4) {
+  console.error('Usage: node script.js directoryPath outputFilePath');
+  process.exit(1);
+}
+
+const directoryPath = process.argv[2];
+const outputFilePath = process.argv[3];
+
+if (!fs.existsSync(directoryPath) || !fs.statSync(directoryPath).isDirectory()) {
+  console.error('Invalid directory path');
+  process.exit(1);
+}
+
 // 現在の作業ディレクトリを取得
 const rootDirectory = process.cwd();
-const fileTree = createTree(rootDirectory);
+const fileTree = createTree(directoryPath);
 
 // JSONに変換
 const jsonTree = JSON.stringify(fileTree, null, 2);
 
-// JSONをファイルに書き出す
-const outputFilePath = path.join(rootDirectory, 'menuThumbnail.json');
+// JSONを指定されたアウトプットファイルに書き出す
 fs.writeFileSync(outputFilePath, jsonTree);
 
 console.log(`File tree JSON data has been written to ${outputFilePath}`);
